dev-requirements.txt
black
django-minio-storage
mypy
coverage
pre-commit